moved tests to a new file
